Fixed at last
I think.
Fixed external commands hanging.
Fixed a few crashes related to trying to access a dir when its actually a file.
Fixed files being made instead of dirs.
Fixed a crash when running setup a second time while LTS is already set up. It was trying to make a dir when one already exists.
Added some clarification on some options.
Made it so if LTS cleanup crashes, it doesnt delete the cleanup file. Mostly useful when maintaining LTS.
Added the ability to only copy native script files. This makes it easier to maintain LTS.
Removed some redundant text.
Disabled the update checker. It causes a crash because there is an issue with configparser. Will be converted to JSON in the future.
Fixed the python folder not being deleted properly on windows.
Made it so python doesnt get downloaded a second time when LTS is already set up on windows.
Removed some legacy code (checkupdates in commands).
Removed some unused imports.

+++ runtime/cleanup.py
@@ -4,6 +4,7 @@ import sys
 import time
 import configparser
 import platform
+import traceback
 
 
 class Cleanup:
@@ -33,6 +34,7 @@ class Cleanup:
         :return:
         """
         print("> Welcome to the LTS cleanup script!")
+        print("> This script will delete your workspace and set most of it to factory defaults.")
         print("> Are you sure you want to clean up your workspace? [y/N]")
         inp = input(": ")
         if inp != "y":
@@ -91,7 +93,7 @@ class Cleanup:
         print("> Deleting non-default config...")
         conftime = time.time()
         if os.path.exists(self.confdir):
-            if os.path.exists(os.path.join(self.confdir, "patches")):
+            if os.path.exists(os.path.join(self.confdir, "patches")) and os.path.isdir(os.path.join(self.confdir, "patches")):
                 shutil.rmtree(os.path.join(self.confdir, "patches"))
             for file in os.listdir(self.confdir):
                 if os.path.isfile(os.path.join(self.confdir, file)) and file not in ["mcp.cfg", "version.cfg"]:
@@ -101,7 +103,8 @@ class Cleanup:
         print("> Deleting system specific files from root...")
         systime = time.time()
         for file in ["decompile", "recompile", "reobfuscate", "startclient", "startserver", "updatemcp", "updatemd5"]:
-            os.unlink(file + "." + self.systemext)
+            if os.path.exists(file + "." + self.systemext) and os.path.isfile(file + "." + self.systemext):
+                os.unlink(file + "." + self.systemext)
         print('> Done in %.2f seconds' % (time.time() - systime))
 
         print("> Done!")
@@ -123,4 +126,8 @@ class Cleanup:
 
 if __name__ == '__main__':
     cleanup = Cleanup()
-    cleanup.start()
+    try:
+        cleanup.start()
+    except:
+        traceback.print_exc()
+        sys.exit(1)
