+++ runtime/linux_scripts/cleanup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 ./runtime/bin/python/bin/pypy runtime/cleanup.py "$@"
 
-rm -rf runtime/bin/python
-rm -f ./cleanup.sh
+if [ $? -ne 1 ]; then
+    rm -rf runtime/bin/python
+    rm -f ./cleanup.sh
+fi
